requirements.txt
asgiref==3.4.1
astroid==2.6.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.8.0
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.12.4
idna==3.2
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
openapi-codec==1.3.2
pycparser==2.20
PyJWT==2.1.0
pylint==2.9.6
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
simplejson==3.17.3
six==1.9.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
